addressing grafana test failures on cygnus
grafana container started up quite slowly on cygnus (probably due to old
hardware). This patch increases wait time after `docker run` to allow
the process to initialize properly. This patch also modifies the tests
to exit with an error if dsosds could not successfully be added to
the grafana server.