Fix '-e' in ldmsd.sh in ppc64le
The 'echo' command is taken as shell built-in command that may or may
not take '-e' parameter. In amd64 and arm64 build environment, the '-e'
was not a problem, but in ppc64le environment, the '-e' was not taken by
the echo command. Instead, the '-e' was taken as a string to echo. This
patch modify the affected Dockerfiles to use '/bin/echo' program
instead.